% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_flextable_full.R
\name{add_total_col}
\alias{add_total_col}
\title{Add Total Column to a Data Frame}
\usage{
add_total_col(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame with a `.total_col` column.
}
\description{
Computes row-wise totals of numeric columns and adds a `.total_col` column.
}
