% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_flextable_full.R
\name{add_total_row}
\alias{add_total_row}
\title{Add Total Row to a Data Frame}
\usage{
add_total_row(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame with an additional summary row.
}
\description{
Sums numeric columns across all rows and appends the result as a "Total" row.
}
