% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_flextable.R
\name{add_totals}
\alias{add_totals}
\title{Add Total Calculations to Data}
\usage{
add_totals(
  template_data,
  total_row = FALSE,
  total_col = FALSE,
  filter_condition = NULL,
  total_row_in_data = NULL
)
}
\arguments{
\item{template_data}{Data frame to add totals to}

\item{total_row}{Logical, whether to add a total row}

\item{total_col}{Logical, whether to add a total column}

\item{filter_condition}{Optional filter condition for total row calculation}
}
\value{
Modified data frame with totals added
}
\description{
Helper function to add row and/or column totals to template data.
}
