% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_excel_tables.R
\name{create_excel_template}
\alias{create_excel_template}
\title{Create Excel Template (Flat Header)}
\usage{
create_excel_template(
  template_data,
  sheet,
  dep,
  amt,
  table_name,
  path = "excel_template"
)
}
\arguments{
\item{template_data}{A `data.frame` to be written into the Excel sheet.}

\item{sheet}{A character string naming the Excel sheet.}

\item{dep}{A string specifying the department name, used to organize file structure.}

\item{amt}{A string specifying the Amt name, used to organize file structure.}

\item{table_name}{The name of the table to be used as the file name.}

\item{path}{Base path where the Excel template should be saved. Defaults to `"excel_template"`.}
}
\value{
Writes an `.xlsx` file to disk. Returns no value.
}
\description{
Generates an Excel file from a flat (non-nested) `data.frame`, including formatted headers, subheaders, and data styles.
The output file is saved in a nested directory structure based on department and amount.
}
