% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_flextable.R
\name{create_flextable_grgeko}
\alias{create_flextable_grgeko}
\title{Create Flextable for GRGEKO Format}
\usage{
create_flextable_grgeko(template_data)
}
\arguments{
\item{template_data}{A data frame to be formatted}
}
\value{
A formatted flextable object
}
\description{
Creates a specialized flextable format for GRGEKO reports.
}
