% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_flextable_full.R
\name{create_subsections}
\alias{create_subsections}
\title{Create Subsections with Optional Subtotals}
\usage{
create_subsections(df, subtotals_in_data, subtotals)
}
\arguments{
\item{df}{A data frame with columns `group`, `label`, and value columns (already in wide format).}

\item{subtotals_in_data}{Logical, indicating if subtotal rows are already included in the data.}

\item{subtotals}{Logical, indicating if subtotals should be shown.}
}
\value{
A data frame with `.label` and `total_col` columns for use in `flextable`, including subtotal rows if specified.
}
\description{
Adds nested row groups and optionally calculates subtotals based on grouping variables.
}
