% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_flextable.R, R/fct_flextable_full.R
\name{fat_border}
\alias{fat_border}
\title{Create Fat Border Style}
\usage{
fat_border()

fat_border()
}
\value{
An officer border object with thick black line

An officer border object with thick black line
}
\description{
Creates a thick black border for flextable formatting.

Creates a thick black border for flextable formatting.
}
