% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_flextable_full.R
\name{format_general}
\alias{format_general}
\title{Apply General Formatting to a Flextable}
\usage{
format_general(ft, fontsize = 9)
}
\arguments{
\item{ft}{A `flextable` object to be styled.}

\item{fontsize}{Numeric font size to apply to the whole table (default is 9).}
}
\value{
A styled `flextable` object.
}
\description{
Applies consistent styling to numeric and character columns of a flextable, including padding, alignment,
font settings, and header styling. It also formats nested headers and sets table properties.
}
