% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_flextable_full.R
\name{format_total_col}
\alias{format_total_col}
\title{Format Total Column in a Flextable}
\usage{
format_total_col(ft, col_index)
}
\arguments{
\item{ft}{A `flextable` object.}

\item{col_index}{Integer index or indices of the column(s) to format.}
}
\value{
A `flextable` object with formatted total column.
}
\description{
Applies styling to the total column in a flextable, including bold text and a background color.
}
