% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_flextable_full.R
\name{format_total_rows}
\alias{format_total_rows}
\title{Format Total Rows in a Flextable}
\usage{
format_total_rows(ft, row_index, style = "bold")
}
\arguments{
\item{ft}{A `flextable` object.}

\item{row_index}{Integer vector indicating which rows to format.}

\item{style}{A character string specifying the formatting style. Options are `"bold"`, `"italic"`, or `"none"`.}
}
\value{
A `flextable` object with formatted total rows.
}
\description{
Applies formatting to designated total rows in a flextable, including bolding or italicizing and drawing horizontal lines.
}
