% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_flextable.R
\name{get_unnested_cols}
\alias{get_unnested_cols}
\title{Identify Unnested Columns in Flextable Header}
\usage{
get_unnested_cols(ft)
}
\arguments{
\item{ft}{A `flextable` object with a nested header structure.}
}
\value{
A numeric vector of column indices that are unnested.
}
\description{
Extracts indices of columns in a `flextable` object that do not contain hierarchical headers.
}
