% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{init_body}
\alias{init_body}
\title{Initialize Dashboard Body}
\usage{
init_body(db_content)
}
\arguments{
\item{db_content}{UI components to be displayed in the body section of the dashboard.}
}
\value{
A `bs4Dash::dashboardBody` object.
}
\description{
Creates the body section of a `bs4Dash` dashboard. This function sets padding, includes custom CSS,
initializes `shinyjs` and `shinybrowser`, and adds JavaScript for user tracking and click event handling.
}
