% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{init_header}
\alias{init_header}
\title{Initialize Dashboard Header}
\usage{
init_header(dashboard_title, reference = "https://statistik.tg.ch")
}
\arguments{
\item{dashboard_title}{A string specifying the title to be displayed in the dashboard header.}

\item{reference}{A string URL pointing to the website to be linked from the brand logo. Defaults to Thurgau Statistics homepage.}
}
\value{
A `bs4Dash::dashboardHeader` object to be used in a Shiny UI layout.
}
\description{
Creates a customized `bs4Dash` dashboard header with a brand title and logo linking to a reference URL.
}
