% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{init_sidebar}
\alias{init_sidebar}
\title{Initialize Dashboard Sidebar}
\usage{
init_sidebar(sidebar_content)
}
\arguments{
\item{sidebar_content}{UI elements to be included within the sidebar, such as menu items or input controls.}
}
\value{
A `bs4Dash::dashboardSidebar` object.
}
\description{
Constructs the sidebar for a `bs4Dash` dashboard using the provided content. Sidebar is not minimized by default.
}
