% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_template_data.R
\name{inverse_locf}
\alias{inverse_locf}
\title{Inverse Last Observation Carried Forward (LOCF)}
\usage{
inverse_locf(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A vector with consecutive duplicates replaced by NA.
}
\description{
Replaces consecutive duplicates with NA, retaining only the first unique value.
}
