% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableview.R
\name{mod_tableview_server}
\alias{mod_tableview_server}
\title{Table View Server Module}
\usage{
mod_tableview_server(id, input_values, nested_data)
}
\arguments{
\item{id}{A unique identifier for the module namespace.}

\item{input_values}{A reactive list of user inputs required to fetch and display data
(`dept`, `amt`, `table`, `year`, `download_id`).}

\item{nested_data}{A reactive object representing a nested list structure of data.}
}
\description{
Manages server-side logic for the table viewing module. It handles:
- Selection and rendering of a specific dataset from nested input
- Rendering of a `flextable` and a `DT` table
- Download functionality for the dataset in multiple formats
}
