% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_template_data.R
\name{reverse_template_colname_wide_grouped}
\alias{reverse_template_colname_wide_grouped}
\title{Reverse Wide Grouped Template to Long Format}
\usage{
reverse_template_colname_wide_grouped(
  data,
  group_col = "group",
  label_col = "label",
  year_col = "jahr"
)
}
\arguments{
\item{data}{A wide-format data frame.}

\item{group_col}{Group column name (default is `"group"`).}

\item{label_col}{Label column name (default is `"label"`).}

\item{year_col}{Name for the added year column (default is `"jahr"`).}
}
\value{
A long-format data frame.
}
\description{
Converts data from wide back to long format using group and label identifiers.
}
