% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_template_data.R
\name{reverse_template_nested_header}
\alias{reverse_template_nested_header}
\title{Reverse Nested Header Template}
\usage{
reverse_template_nested_header(data, id_col = "id", year_col = "jahr")
}
\arguments{
\item{data}{A wide-format data frame with nested headers.}

\item{id_col}{Identifier column (default is `"id"`).}

\item{year_col}{Year column to add (default is `"jahr"`).}
}
\value{
A long-format data frame with separated levels and values.
}
\description{
Splits combined nested headers back into hierarchical columns.
}
