% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_template_data.R
\name{template_colname_wide_grouped}
\alias{template_colname_wide_grouped}
\title{Transform Grouped Data from Long to Wide Format with Aggregation}
\usage{
template_colname_wide_grouped(
  data,
  group_col,
  label_col,
  colnames_col,
  val_col,
  year_col = "jahr"
)
}
\arguments{
\item{data}{A data frame in long format.}

\item{group_col}{Column identifying the group variable.}

\item{label_col}{Column identifying subgroup labels.}

\item{colnames_col}{Column containing variable names to pivot wider.}

\item{val_col}{Column containing values to fill.}

\item{year_col}{Column indicating the year (default is `"jahr"`).}
}
\value{
A wide-format data frame with aggregated subgroups and labels.
}
\description{
Converts a long-format data frame with grouping columns into a wide format.
Also aggregates and relabels subgroups under the same group. (for Example Verwaltungsrechtspflege)
}
