% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_flextable_full.R
\name{template_double_col_no_header}
\alias{template_double_col_no_header}
\title{Create Nested Table with No Header and Two Column Levels}
\usage{
template_double_col_no_header(data, column_cols)
}
\arguments{
\item{data}{A data frame in long format.}

\item{column_cols}{A character vector of length 2 indicating group and label columns.}
}
\value{
A renamed data frame.
}
\description{
Prepares a dataset using group and label columns without header information.
}
