% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_flextable_full.R
\name{template_double_col_single_header}
\alias{template_double_col_single_header}
\title{Create Nested Table with Single-Level Header}
\usage{
template_double_col_single_header(data, header_cols, column_cols, val_col)
}
\arguments{
\item{data}{A data frame in long format.}

\item{header_cols}{A character vector of length 1 indicating the header column (e.g., `"indikator"`).}

\item{column_cols}{A character vector of length 2 indicating the row group and label columns.}

\item{val_col}{A string naming the column containing the numeric values.}
}
\value{
A renamed data frame
}
\description{
Transforms and organizes data to prepare a single-level column header and nested row labels with optional subtotals.
}
