% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_template_data.R
\name{template_nested_header}
\alias{template_nested_header}
\title{Transform Data into Nested Header Format}
\usage{
template_nested_header(
  data,
  level1_col,
  level2_col,
  id_col,
  val_col,
  year_col = "jahr"
)
}
\arguments{
\item{data}{A data frame.}

\item{level1_col}{Top-level header column.}

\item{level2_col}{Sub-level header column.}

\item{id_col}{Identifier column.}

\item{val_col}{Value column.}

\item{year_col}{Column indicating the year (default is `"jahr"`).}
}
\value{
A data frame in wide format with nested headers.
}
\description{
Combines two categorical levels (`level1`, `level2`) into a nested column header format for wide transformation.
}
