% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_flextable_full.R
\name{template_no_col_single_header}
\alias{template_no_col_single_header}
\title{Create Table with No Column Grouping and Single Header}
\usage{
template_no_col_single_header(data, header_cols, column_cols, val_col)
}
\arguments{
\item{data}{A data frame.}

\item{header_cols}{A character vector of length 1.}

\item{column_cols}{A character vector of length 1 (used for labeling).}

\item{val_col}{A string for the value column.}
}
\value{
A renamed data frame.
}
\description{
Useful for simple tables with one header and no explicit grouping columns.
}
