% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_flextable_full.R
\name{template_single_col_double_header}
\alias{template_single_col_double_header}
\title{Create Nested Table with One Column and Two Header Levels}
\usage{
template_single_col_double_header(data, header_cols, column_cols, val_col)
}
\arguments{
\item{data}{A data frame in long format.}

\item{header_cols}{A character vector of length 2 for the header columns.}

\item{column_cols}{A character vector of length 1 for the row label.}

\item{val_col}{A string naming the column containing the numeric values.}
}
\value{
A renamed data frame.
}
\description{
Transforms and organizes data to include a double-level header and a single row label.
}
