% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_flextable_full.R
\name{template_single_col_single_header}
\alias{template_single_col_single_header}
\title{Create Nested Table with One Column and One Header}
\usage{
template_single_col_single_header(data, header_cols, column_cols, val_col)
}
\arguments{
\item{data}{A data frame.}

\item{header_cols}{A character vector of length 1.}

\item{column_cols}{A character vector of length 1.}

\item{val_col}{A string for the value column name.}
}
\value{
A renamed data frame.
}
\description{
Transforms data with one column and one header into a renamed format.
}
