% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_flextable.R, R/fct_flextable_full.R
\name{thin_border}
\alias{thin_border}
\title{Create Thin Border Style}
\usage{
thin_border()

thin_border()
}
\value{
An officer border object with thin black line

An officer border object with thin black line
}
\description{
Creates a thin black border for flextable formatting.

Creates a thin black border for flextable formatting.
}
