% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_flextable.R
\name{produce_flextable}
\alias{produce_flextable}
\title{Generate Flextable Based on Table Type}
\usage{
produce_flextable(data_list, dep, amt, table, year)
}
\arguments{
\item{data_list}{A nested list of elements containing table metadata and data.}

\item{dep}{Department key in `data_list`.}

\item{amt}{Amt key in `data_list`.}

\item{table}{Table key in `data_list`.}

\item{year}{Year to filter the data on.}
}
\value{
A `flextable` object with custom styling depending on `table_type`.
}
\description{
Dispatches to different formatting and transformation routines based on the `table_type` in `elem`.
Supports: `"nested_header"`, `"none"`, `"nested_col"`, `"colname_wide_grouped"`.
}
