% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_template_data.R
\name{template_nested_col}
\alias{template_nested_col}
\title{Transform to Nested Columns with Level Names}
\usage{
template_nested_col(
  data,
  group_col,
  label_col,
  val_col,
  year_col = "jahr",
  num_years,
  colnames_col,
  no_subtotals = F,
  new_names,
  subtotals_in_data = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{val_col}{Value column.}

\item{year_col}{Year column (default is `"jahr"`).}

\item{colnames_col}{Column containing variable names.}

\item{level1_col}{Higher-level grouping column.}

\item{level2_col}{Sub-level column.}

\item{level1_name}{Optional name override for `level1_col`.}

\item{level2_name}{Optional name override for `level2_col`.}
}
\value{
A wide-format data frame with nested columns.
}
\description{
Combines hierarchical levels and spreads the data into a wide format. Applies inverse LOCF to higher level.
}
